requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.11.0.86
boto3==1.38.1
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pandas==2.2.3
numpy==2.2.5
starlette==0.27.0
typing-extensions==4.13.2
scikit-learn==1.6.1
Pillow==10.2.0
firebase-admin==6.4.0
pytest==8.3.5
pytest-asyncio==0.23.5
httpx==0.27.0
sqlalchemy_utils==0.41.1 